refactor(shared-server): split topology mutation core
