+++ packages/shared-server/rallar-system/topology/config/mutation/validate-topology-config-receipt.ts
@@ -1,0 +1,236 @@
+// prettier-ignore
+import type {
+  GroupTopologyConfigMutationReceipt,
+} from '@shared/api/graph-topology-management-types.ts';
+import type { GroupRef, GroupStateCausalRevision } from '@shared/api/group-types.ts';
+import {
+  isTopologyConfigRecord,
+  requireTopologyString,
+  sameTopologyGroupRef,
+  validateAcceptedTopologyConfig,
+  validateTopologyConfigExactKeys,
+  validateTopologyGroupRef,
+  validateTopologyPositiveInteger,
+  validateTopologyStorageRevision,
+} from './topology-config-mutation-validation-primitives.ts';
+
+export function validateTopologyConfigReceipt(
+  value: unknown,
+  expectedRef: GroupRef,
+): asserts value is GroupTopologyConfigMutationReceipt {
+  if (!isTopologyConfigRecord(value)) throw new TypeError('Topology config receipt is invalid');
+  validateTopologyConfigReceiptIdentity(value, expectedRef);
+  validateTopologyConfigReceiptAcceptedState(value);
+  validateTopologyConfigReceiptEffect(value);
+  validateTopologyConfigReceiptCausalRevision(value);
+  validateTopologyConfigReceiptTimestamps(value);
+}
+
+function validateTopologyConfigReceiptIdentity(
+  receipt: Readonly<Record<string, unknown>>,
+  expectedRef: GroupRef,
+): void {
+  validateTopologyConfigExactKeys(receipt, topologyConfigReceiptKeys, 'Topology config receipt');
+  requireTopologyString(receipt.commandId, 'Topology config receipt commandId');
+  if (receipt.requestId !== null) {
+    requireTopologyString(receipt.requestId, 'Topology config receipt requestId');
+  }
+  validateTopologyPositiveInteger(receipt.attemptCount, 'Topology config receipt attemptCount');
+  if (!/^sha256:[0-9a-f]{64}$/.test(String(receipt.commandHash))) {
+    throw new TypeError('Topology config receipt hash is invalid');
+  }
+  if (!topologyConfigOperations.includes(String(receipt.operation))) {
+    throw new TypeError('Topology config receipt operation is invalid');
+  }
+  if (receipt.outcome !== 'applied' && receipt.outcome !== 'no-op') {
+    throw new TypeError('Topology config receipt outcome is invalid');
+  }
+  validateTopologyGroupRef(receipt.groupRef, 'Topology config receipt groupRef');
+  if (!sameTopologyGroupRef(receipt.groupRef as GroupRef, expectedRef)) {
+    throw new TypeError('Topology config receipt has the wrong groupRef');
+  }
+  if (receipt.target !== 'config' && receipt.target !== 'override') {
+    throw new TypeError('Topology config receipt target is invalid');
+  }
+}
+
+function validateTopologyConfigReceiptAcceptedState(
+  receipt: Readonly<Record<string, unknown>>,
+): void {
+  const expectsConfig = receipt.operation === 'putConfig' || receipt.operation === 'deleteConfig';
+  if ((expectsConfig ? 'config' : 'override') !== receipt.target) {
+    throw new TypeError('Topology config receipt operation target is invalid');
+  }
+  const isPut = receipt.operation === 'putConfig' || receipt.operation === 'putOverride';
+  if (isPut && receipt.outcome !== 'applied') {
+    throw new TypeError('Topology config PUT receipt must be applied');
+  }
+  validateTopologyStorageRevision(receipt.acceptedVersion, 'Topology config accepted version');
+  if (receipt.acceptedStorageRevision !== null) {
+    validateTopologyStorageRevision(
+      receipt.acceptedStorageRevision,
+      'Topology config accepted storage revision',
+    );
+  }
+  for (const [field, label] of acceptedReceiptTimeFields) {
+    if (receipt[field] !== null) validateTopologyStorageRevision(receipt[field], label);
+  }
+  if (receipt.acceptedConfig !== null) {
+    validateAcceptedTopologyConfig(
+      receipt.acceptedConfig,
+      'Topology config receipt accepted config',
+    );
+  }
+  if (isPut !== (receipt.acceptedConfig !== null)) {
+    throw new TypeError('Topology config receipt accepted config does not match operation');
+  }
+}
+
+function validateTopologyConfigReceiptEffect(receipt: Readonly<Record<string, unknown>>): void {
+  if (receipt.outboxId !== null) {
+    requireTopologyString(receipt.outboxId, 'Topology config outboxId');
+  }
+  if (receipt.eventId !== null) throw new TypeError('Topology config receipt eventId must be null');
+  if (
+    !Array.isArray(receipt.outboxIds) ||
+    receipt.outboxIds.some((outboxId) => typeof outboxId !== 'string') ||
+    receipt.outboxIds.length !== (receipt.outboxId === null ? 0 : 1) ||
+    (receipt.outboxId !== null && receipt.outboxIds[0] !== receipt.outboxId)
+  ) {
+    throw new TypeError('Topology config receipt outboxIds are invalid');
+  }
+  if (
+    receipt.outcome === 'applied' &&
+    (Number(receipt.acceptedVersion) <= 0 ||
+      receipt.acceptedStorageRevision === null ||
+      receipt.acceptedCausalRevision === null ||
+      receipt.outboxId === null)
+  ) {
+    throw new TypeError('Topology config applied receipt is incomplete');
+  }
+  if (
+    (receipt.outcome === 'applied') !== (receipt.outboxId !== null) ||
+    (receipt.outcome === 'applied') !== (receipt.acceptedCausalRevision !== null)
+  ) {
+    throw new TypeError('Topology config receipt effect does not match outboxId');
+  }
+}
+
+function validateTopologyConfigReceiptCausalRevision(
+  receipt: Readonly<Record<string, unknown>>,
+): void {
+  if (receipt.acceptedCausalRevision === null) return;
+  if (!isTopologyConfigRecord(receipt.acceptedCausalRevision)) {
+    throw new TypeError('Topology config accepted causal revision is invalid');
+  }
+  const accepted = receipt.acceptedCausalRevision;
+  validateTopologyConfigExactKeys(
+    accepted,
+    topologyConfigAcceptedCausalRevisionKeys,
+    'Topology config accepted causal revision',
+  );
+  for (const field of causalReceiptRevisionFields) {
+    validateTopologyStorageRevision(
+      accepted[field],
+      `Topology config accepted causal revision ${field}`,
+    );
+  }
+  validateTopologyCausalRevision(
+    accepted.causalRevision,
+    'Topology config accepted causal revision tuple',
+  );
+  const causalRevision = accepted.causalRevision as GroupStateCausalRevision;
+  const expectedOutboxId = [
+    String(receipt.commandId),
+    'rtc-topology-recompute',
+    'group-revision',
+    `group=${causalRevision.groupRevision};presence=${causalRevision.presenceRevision}`,
+  ].join(':');
+  if (receipt.outboxId !== expectedOutboxId) {
+    throw new TypeError('Topology config receipt outbox identity is invalid');
+  }
+}
+
+function validateTopologyConfigReceiptTimestamps(receipt: Readonly<Record<string, unknown>>): void {
+  const isPut = receipt.operation === 'putConfig' || receipt.operation === 'putOverride';
+  if (
+    receipt.outcome === 'no-op' &&
+    Number(receipt.acceptedVersion) === 0 &&
+    receipt.acceptedStorageRevision !== null
+  ) {
+    throw new TypeError('Topology config absent no-op receipt is invalid');
+  }
+  if (
+    isPut !==
+    (receipt.acceptedCreatedAtEpochMs !== null && receipt.acceptedUpdatedAtEpochMs !== null)
+  ) {
+    throw new TypeError('Topology config receipt timestamps do not match operation');
+  }
+  if (
+    receipt.acceptedCreatedAtEpochMs !== null &&
+    Number(receipt.acceptedUpdatedAtEpochMs) < Number(receipt.acceptedCreatedAtEpochMs)
+  ) {
+    throw new TypeError('Topology config receipt update precedes creation');
+  }
+  if ((receipt.operation === 'putOverride') !== (receipt.acceptedExpiresAtEpochMs !== null)) {
+    throw new TypeError('Topology config receipt expiry does not match operation');
+  }
+  if (
+    receipt.acceptedExpiresAtEpochMs !== null &&
+    Number(receipt.acceptedExpiresAtEpochMs) <= Number(receipt.acceptedUpdatedAtEpochMs)
+  ) {
+    throw new TypeError('Topology config receipt expiry does not follow update');
+  }
+}
+
+function validateTopologyCausalRevision(
+  value: unknown,
+  label: string,
+): asserts value is GroupStateCausalRevision {
+  if (!isTopologyConfigRecord(value)) throw new TypeError(`${label} is invalid`);
+  validateTopologyConfigExactKeys(value, ['groupRevision', 'presenceRevision'], label);
+  validateTopologyStorageRevision(value.groupRevision, `${label} groupRevision`);
+  validateTopologyStorageRevision(value.presenceRevision, `${label} presenceRevision`);
+}
+
+const topologyConfigOperations = ['putConfig', 'deleteConfig', 'putOverride', 'deleteOverride'];
+const topologyConfigReceiptKeys = [
+  'commandId',
+  'requestId',
+  'commandHash',
+  'operation',
+  'outcome',
+  'attemptCount',
+  'groupRef',
+  'target',
+  'acceptedVersion',
+  'acceptedStorageRevision',
+  'acceptedCreatedAtEpochMs',
+  'acceptedUpdatedAtEpochMs',
+  'acceptedExpiresAtEpochMs',
+  'acceptedConfig',
+  'acceptedCausalRevision',
+  'eventId',
+  'outboxId',
+  'outboxIds',
+];
+const topologyConfigAcceptedCausalRevisionKeys = [
+  'stateRevision',
+  'causalRevision',
+  'snapshotVersion',
+  'metadataVersion',
+  'rosterVersion',
+  'presenceVersion',
+];
+const acceptedReceiptTimeFields = [
+  ['acceptedCreatedAtEpochMs', 'Topology config accepted creation time'],
+  ['acceptedUpdatedAtEpochMs', 'Topology config accepted update time'],
+  ['acceptedExpiresAtEpochMs', 'Topology config accepted expiry'],
+] as const;
+const causalReceiptRevisionFields = [
+  'stateRevision',
+  'snapshotVersion',
+  'metadataVersion',
+  'rosterVersion',
+  'presenceVersion',
+] as const;
