+++ packages/shared-server/rallar-system/topology/inbox/topology-app-inbox-contracts.ts
@@ -8,7 +8,7 @@ import type { GroupTopologyConfigPatch } from '@shared/api/graph-topology-manage
 // prettier-ignore
 import type {
   TopologyMutationAuthorityProof,
-} from '../../services/topology-mutation-authority-proof.ts';
+} from './topology-mutation-authority-proof.ts';
 
 export type TopologyAppInboxOperation =
   | 'putConfig'
@@ -18,57 +18,69 @@ export type TopologyAppInboxOperation =
   | 'reconfigureTopology'
   | 'submitRtt';
 
-export type TopologyAppInboxRequestPayload =
-  | Readonly<{
-      operation: 'putConfig';
-      config: GroupTopologyConfigPatch;
-    }>
-  | Readonly<{
-      operation: 'deleteConfig';
-      target: 'config';
-    }>
-  | Readonly<{
-      operation: 'putOverride';
-      config: GroupTopologyConfigPatch;
-      ttlMs: number | null;
-      expiresAtEpochMs: number | null;
-    }>
-  | Readonly<{
-      operation: 'deleteOverride';
-      target: 'override';
-    }>
-  | Readonly<{
-      operation: 'reconfigureTopology';
-      requestOptions: GroupTopologyConfigPatch;
-      publish: boolean;
-    }>;
+type TopologyAppInboxRequestPayloadByOperation = Readonly<{
+  putConfig: Readonly<{
+    operation: 'putConfig';
+    config: GroupTopologyConfigPatch;
+  }>;
+  deleteConfig: Readonly<{
+    operation: 'deleteConfig';
+    target: 'config';
+  }>;
+  putOverride: Readonly<{
+    operation: 'putOverride';
+    config: GroupTopologyConfigPatch;
+    ttlMs: number | null;
+    expiresAtEpochMs: number | null;
+  }>;
+  deleteOverride: Readonly<{
+    operation: 'deleteOverride';
+    target: 'override';
+  }>;
+  reconfigureTopology: Readonly<{
+    operation: 'reconfigureTopology';
+    requestOptions: GroupTopologyConfigPatch;
+    publish: boolean;
+  }>;
+}>;
 
-export type TopologyAppInboxPayload =
-  | Readonly<{
-      operation: 'putConfig';
-      config: CanonicalGroupTopologyConfigPatch;
-    }>
-  | Readonly<{
-      operation: 'deleteConfig';
-      target: 'config';
-    }>
-  | Readonly<{
-      operation: 'putOverride';
-      config: CanonicalGroupTopologyConfigPatch;
-      ttlMs: number | null;
-      expiresAtEpochMs: number | null;
-    }>
-  | Readonly<{
-      operation: 'deleteOverride';
-      target: 'override';
-    }>
-  | Readonly<{
-      operation: 'reconfigureTopology';
-      requestOptions: CanonicalGroupTopologyConfigPatch;
-      publish: boolean;
-    }>;
+type TopologyAppInboxPayloadByOperation = Readonly<{
+  putConfig: Readonly<{
+    operation: 'putConfig';
+    config: CanonicalGroupTopologyConfigPatch;
+  }>;
+  deleteConfig: Readonly<{
+    operation: 'deleteConfig';
+    target: 'config';
+  }>;
+  putOverride: Readonly<{
+    operation: 'putOverride';
+    config: CanonicalGroupTopologyConfigPatch;
+    ttlMs: number | null;
+    expiresAtEpochMs: number | null;
+  }>;
+  deleteOverride: Readonly<{
+    operation: 'deleteOverride';
+    target: 'override';
+  }>;
+  reconfigureTopology: Readonly<{
+    operation: 'reconfigureTopology';
+    requestOptions: CanonicalGroupTopologyConfigPatch;
+    publish: boolean;
+  }>;
+}>;
 
-export type TopologyAppInboxCommand = Readonly<{
+type TopologyAppInboxCommandOperation = keyof TopologyAppInboxPayloadByOperation;
+
+export type TopologyAppInboxRequestPayload<
+  Operation extends TopologyAppInboxCommandOperation = TopologyAppInboxCommandOperation,
+> = TopologyAppInboxRequestPayloadByOperation[Operation];
+
+export type TopologyAppInboxPayload<
+  Operation extends TopologyAppInboxCommandOperation = TopologyAppInboxCommandOperation,
+> = TopologyAppInboxPayloadByOperation[Operation];
+
+type TopologyAppInboxCommandFor<Operation extends TopologyAppInboxCommandOperation> = Readonly<{
   actor: Readonly<{
     principalId: string;
     sessionId: string;
@@ -77,17 +89,26 @@ export type TopologyAppInboxCommand = Readonly<{
   requestId: string;
   commandHash: string;
   capturedAtEpochMs: number;
-  operation: Exclude<TopologyAppInboxOperation, 'submitRtt'>;
-  payload: TopologyAppInboxPayload;
+  operation: Operation;
+  payload: TopologyAppInboxPayload<Operation>;
 }>;
 
-export type CreateTopologyAppInboxCommandInput = Readonly<{
-  actor: TopologyAppInboxCommand['actor'];
-  groupRef: GroupRef;
-  requestId: string;
-  capturedAtEpochMs: number;
-  payload: TopologyAppInboxRequestPayload;
-}>;
+export type TopologyAppInboxCommand = {
+  [Operation in TopologyAppInboxCommandOperation]: TopologyAppInboxCommandFor<Operation>;
+}[TopologyAppInboxCommandOperation];
+
+type CreateTopologyAppInboxCommandInputFor<Operation extends TopologyAppInboxCommandOperation> =
+  Readonly<{
+    actor: TopologyAppInboxCommand['actor'];
+    groupRef: GroupRef;
+    requestId: string;
+    capturedAtEpochMs: number;
+    payload: TopologyAppInboxRequestPayload<Operation>;
+  }>;
+
+export type CreateTopologyAppInboxCommandInput<
+  Operation extends TopologyAppInboxCommandOperation = TopologyAppInboxCommandOperation,
+> = CreateTopologyAppInboxCommandInputFor<Operation>;
 
 export type TopologyConfigAppInboxAuthority = Readonly<{
   kind: 'topology-config';
