--- packages/tests/shared-server/group-topology-config-service.test.ts
@@ -1,763 +1,0 @@
-import { describe, expect, expectTypeOf, it } from 'vitest';
-import { readFileSync } from 'node:fs';
-import type {
-  PutGroupTopologyConfigRequest,
-  PutGroupTopologyOverrideRequest,
-} from '@shared/api/graph-topology-management-types.ts';
-import type { GroupSnapshot } from '@shared/api/group-types.ts';
-import {
-  DEFAULT_GROUP_TOPOLOGY_OVERRIDE_TTL_MS,
-  GroupTopologyConfigValidationError,
-  MAX_GROUP_TOPOLOGY_OVERRIDE_TTL_MS,
-  readDefaultGroupTopologyConfig,
-  resolveGroupTopologyConfig,
-  resolveOverrideExpiresAtEpochMs,
-  validateEffectiveGroupTopologyConfig,
-  validateGroupTopologyConfigPatch,
-} from '@shared-server/rallar-system/services/group-topology-config-service.ts';
-import {
-  computeTopologyConfigMutation,
-  validateGroupTopologyConfigMutationRecord,
-  validateTopologyConfigMutation,
-} from '@shared-server/rallar-system/services/group-topology-config-mutations.ts';
-
-describe('group topology config service', () => {
-  it('keeps synchronous reconfigure options off config mutation requests', () => {
-    type ConfigHasReconfigure = 'reconfigure' extends keyof PutGroupTopologyConfigRequest
-      ? true
-      : false;
-    type OverrideHasReconfigure = 'reconfigure' extends keyof PutGroupTopologyOverrideRequest
-      ? true
-      : false;
-    expectTypeOf<ConfigHasReconfigure>().toEqualTypeOf<false>();
-    expectTypeOf<OverrideHasReconfigure>().toEqualTypeOf<false>();
-  });
-
-  it('computes and validates the same immutable config mutation twice', () => {
-    const input = deepFreeze({
-      command: {
-        operation: 'putConfig' as const,
-        aggregateRef: createGroupRef(),
-        commandId: 'config-command-1',
-        requestId: 'config-command-1',
-        input: {
-          config: { topologyKind: 'tree' as const, degreeLimit: 4 },
-          updatedByPrincipalId: 'owner',
-          ttlMs: null,
-          expiresAtEpochMs: null,
-        },
-      },
-      read: {
-        config: null,
-        override: null,
-        configGeneration: null,
-        overrideGeneration: null,
-        invariantGeneration: null,
-        idempotency: null,
-        groupSnapshot: createGroupSnapshot(),
-        groupAuthorityGuard: createGroupAuthorityGuard(40),
-      },
-      facts: {
-        requestedAtEpochMs: 1_000,
-        policyNowEpochMs: 1_000,
-        commandHash: `sha256:${'a'.repeat(64)}`,
-        attemptCount: 1,
-        isPlatformAdmin: false,
-        resolvedOverrideExpiresAtEpochMs: null,
-        deleteTarget: null,
-      },
-      serverDefaults: {},
-    });
-    const before = structuredClone(input);
-
-    const first = computeTopologyConfigMutation(input);
-    const second = computeTopologyConfigMutation(input);
-    const laterPolicyInput = deepFreeze({
-      ...input,
-      facts: { ...input.facts, policyNowEpochMs: 2_000 },
-    });
-    const laterPolicy = computeTopologyConfigMutation(laterPolicyInput);
-
-    expect(first).toEqual(second);
-    expect(laterPolicy).toEqual(first);
-    expect(input).toEqual(before);
-    expect(() => validateTopologyConfigMutation({ ...input, computed: first })).not.toThrow();
-    expect(() => validateTopologyConfigMutation({ ...input, computed: second })).not.toThrow();
-    expect(() =>
-      validateTopologyConfigMutation({
-        ...laterPolicyInput,
-        computed: laterPolicy,
-      }),
-    ).not.toThrow();
-
-    if (first.outcome !== 'write') {
-      throw new Error('Expected an applied topology config mutation');
-    }
-    expect(() =>
-      validateGroupTopologyConfigMutationRecord(
-        {
-          groupRef: input.command.aggregateRef,
-          requestId: input.command.requestId,
-          commandHash: input.facts.commandHash,
-          receipt: {
-            ...first.receipt,
-            outboxId: 'state-mutation-attacker-selected',
-          },
-        },
-        {
-          groupRef: input.command.aggregateRef,
-          requestId: input.command.requestId,
-        },
-      ),
-    ).toThrow('Topology config receipt outboxIds are invalid');
-    expect(() =>
-      validateGroupTopologyConfigMutationRecord(
-        {
-          groupRef: input.command.aggregateRef,
-          requestId: input.command.requestId,
-          commandHash: input.facts.commandHash,
-          receipt: { ...first.receipt, acceptedConfig: null },
-        },
-        {
-          groupRef: input.command.aggregateRef,
-          requestId: input.command.requestId,
-        },
-      ),
-    ).toThrow('accepted config does not match operation');
-    expect(() =>
-      validateGroupTopologyConfigMutationRecord(
-        {
-          groupRef: input.command.aggregateRef,
-          requestId: input.command.requestId,
-          commandHash: input.facts.commandHash,
-          receipt: {
-            ...first.receipt,
-            acceptedConfig: { topologyKind: 'tree' },
-          },
-        },
-        {
-          groupRef: input.command.aggregateRef,
-          requestId: input.command.requestId,
-        },
-      ),
-    ).toThrow('accepted config fields are invalid');
-  });
-
-  it('clears durable and override fields back to their immediate fallback', () => {
-    const durableInput = createConfigMutationInput({
-      operation: 'putConfig',
-      config: { degreeLimit: null },
-      durableDegreeLimit: 9,
-      overrideDegreeLimit: null,
-    });
-    const durable = computeTopologyConfigMutation(durableInput);
-    if (durable.outcome !== 'write' || durable.result.kind !== 'config') {
-      throw new Error('Expected durable config write');
-    }
-    expect(durable.result.config.config.degreeLimit).toBe(5);
-
-    const overrideInput = createConfigMutationInput({
-      operation: 'putOverride',
-      config: { degreeLimit: null },
-      durableDegreeLimit: 4,
-      overrideDegreeLimit: 9,
-    });
-    const override = computeTopologyConfigMutation(overrideInput);
-    if (override.outcome !== 'write' || override.result.kind !== 'override') {
-      throw new Error('Expected topology override write');
-    }
-    expect(override.result.override.config.degreeLimit).toBe(4);
-  });
-
-  it('rejects an invalid durable config even when a temporary override hides it until expiry', () => {
-    const input = createConfigMutationInput({
-      operation: 'putConfig',
-      config: { meshParamK: 4 },
-      durableDegreeLimit: 3,
-      overrideDegreeLimit: 5,
-    });
-
-    expect(() =>
-      computeTopologyConfigMutation({
-        ...input,
-        serverDefaults: { degreeLimit: 3, meshParamK: 2 },
-      }),
-    ).toThrow(GroupTopologyConfigValidationError);
-  });
-
-  it('denies expired and terminal lifecycle mutations to platform admins', () => {
-    const input = createConfigMutationInput({
-      operation: 'putConfig',
-      config: { topologyKind: 'tree' },
-      durableDegreeLimit: 5,
-      overrideDegreeLimit: null,
-    });
-    const expired = {
-      ...input.read.groupSnapshot,
-      group: {
-        ...input.read.groupSnapshot.group,
-        expiresAtEpochMs: 1_500,
-      },
-    };
-    const terminal = {
-      ...input.read.groupSnapshot,
-      group: {
-        ...input.read.groupSnapshot.group,
-        status: 'deleted' as const,
-      },
-    };
-
-    for (const [groupSnapshot, denialCode] of [
-      [expired, 'group-not-active'],
-      [terminal, 'group-deleted'],
-    ] as const) {
-      expect(() =>
-        computeTopologyConfigMutation({
-          ...input,
-          read: { ...input.read, groupSnapshot },
-          facts: {
-            ...input.facts,
-            isPlatformAdmin: true,
-            policyNowEpochMs: 2_000,
-          },
-        }),
-      ).toThrow(
-        expect.objectContaining({
-          status: 403,
-          denial: expect.objectContaining({ code: denialCode }),
-        }),
-      );
-    }
-  });
-
-  it('rejects compact replay receipt operation corruption against the verified command', () => {
-    const input = createConfigMutationInput({
-      operation: 'putConfig',
-      config: { topologyKind: 'tree' },
-      durableDegreeLimit: 5,
-      overrideDegreeLimit: null,
-    });
-    const accepted = computeTopologyConfigMutation(input);
-    if (accepted.outcome !== 'write') {
-      throw new Error('Expected topology config write');
-    }
-    const corruptRecord = {
-      groupRef: input.command.aggregateRef,
-      requestId: input.command.requestId,
-      commandHash: input.facts.commandHash,
-      receipt: {
-        ...accepted.receipt,
-        operation: 'putOverride' as const,
-        target: 'override' as const,
-        acceptedExpiresAtEpochMs: accepted.receipt.acceptedUpdatedAtEpochMs! + 1,
-      },
-    };
-    const replayInput = {
-      ...input,
-      read: {
-        ...input.read,
-        idempotency: {
-          key: 'corrupt-replay',
-          value: corruptRecord,
-          entry: {
-            key: 'corrupt-replay',
-            value: JSON.stringify(corruptRecord),
-            expireAtTimestamp: Number.MAX_SAFE_INTEGER,
-            updatedTimestamp: new Date(0).toISOString(),
-            revision: 0,
-          },
-        },
-      },
-    };
-
-    expect(() => computeTopologyConfigMutation(replayInput)).toThrow(
-      'Topology config receipt operation differs from command',
-    );
-  });
-
-  it.each(['putConfig', 'putOverride'] as const)(
-    'rejects an impossible %s no-op receipt at the pure validator boundary',
-    (operation) => {
-      const groupRef = createGroupRef();
-      const requestId = `impossible-${operation}`;
-      const commandHash = `sha256:${'7'.repeat(64)}`;
-      expect(() =>
-        validateGroupTopologyConfigMutationRecord(
-          {
-            groupRef,
-            requestId,
-            commandHash,
-            receipt: {
-              commandId: requestId,
-              requestId,
-              commandHash,
-              operation,
-              outcome: 'no-op',
-              attemptCount: 1,
-              groupRef,
-              target: operation === 'putConfig' ? 'config' : 'override',
-              acceptedVersion: 1,
-              acceptedStorageRevision: null,
-              acceptedCreatedAtEpochMs: 1_000,
-              acceptedUpdatedAtEpochMs: 1_000,
-              acceptedExpiresAtEpochMs: operation === 'putOverride' ? 6_000 : null,
-              acceptedConfig: {
-                topologyKind: 'tree',
-                degreeLimit: 5,
-                treeMinSize: 5,
-                meshMinSize: 16,
-                meshParamK: 2,
-              },
-              acceptedCausalRevision: null,
-              eventId: null,
-              outboxId: null,
-              outboxIds: [],
-            },
-          },
-          { groupRef, requestId },
-        ),
-      ).toThrow('Topology config PUT receipt must be applied');
-    },
-  );
-
-  it('rejects an elapsed stable override expiry with explicit pure facts', () => {
-    const input = deepFreeze({
-      command: {
-        operation: 'putOverride' as const,
-        aggregateRef: createGroupRef(),
-        commandId: 'elapsed-stable-expiry',
-        requestId: 'elapsed-stable-expiry',
-        input: {
-          config: { topologyKind: 'tree' as const },
-          updatedByPrincipalId: 'owner',
-          ttlMs: 5_000,
-          expiresAtEpochMs: null,
-        },
-      },
-      read: {
-        config: null,
-        override: null,
-        configGeneration: null,
-        overrideGeneration: null,
-        invariantGeneration: null,
-        idempotency: null,
-        groupSnapshot: createGroupSnapshot(),
-        groupAuthorityGuard: createGroupAuthorityGuard(),
-      },
-      facts: {
-        requestedAtEpochMs: 1_000,
-        policyNowEpochMs: 7_000,
-        commandHash: `sha256:${'8'.repeat(64)}`,
-        attemptCount: 1,
-        isPlatformAdmin: false,
-        resolvedOverrideExpiresAtEpochMs: 6_000,
-        deleteTarget: null,
-      },
-      serverDefaults: {},
-    });
-
-    expect(() => computeTopologyConfigMutation(input)).toThrow(GroupTopologyConfigValidationError);
-  });
-
-  it('keeps pure topology config phases ambient-free and orchestration visible', () => {
-    const mutationSource = readFileSync(
-      new URL(
-        '../../shared-server/rallar-system/services/group-topology-config-mutations.ts',
-        import.meta.url,
-      ),
-      'utf8',
-    );
-    for (const forbidden of [
-      'Date.now',
-      'Temporal.Now',
-      'Math.random',
-      'randomUUID',
-      '.begin(',
-      'hashMutationCommand',
-      'publisher',
-      'topologyService',
-    ]) {
-      expect(mutationSource, forbidden).not.toContain(forbidden);
-    }
-
-    const managementSource = readFileSync(
-      new URL(
-        '../../shared-server/rallar-system/services/group-topology-management-service.ts',
-        import.meta.url,
-      ),
-      'utf8',
-    );
-    const appInboxSource = readFileSync(
-      new URL(
-        '../../shared-server/rallar-system/topology/inbox/topology-app-inbox-handler.ts',
-        import.meta.url,
-      ),
-      'utf8',
-    );
-    const read = appInboxSource.indexOf(
-      'const read = await topologyManagementService.readTopologyConfigMutation',
-    );
-    const compute = appInboxSource.indexOf(
-      'const computed = topologyManagementService.computeTopologyConfigMutation',
-      read,
-    );
-    const validate = appInboxSource.indexOf(
-      'topologyManagementService.validateTopologyConfigMutation',
-      compute,
-    );
-    const transaction = appInboxSource.indexOf(
-      'const result = await this.dependencies.writeMutation',
-      validate,
-    );
-    const write = appInboxSource.indexOf(
-      'await topologyManagementService.writeTopologyConfigMutation',
-      transaction,
-    );
-    expect(read).toBeGreaterThan(-1);
-    expect(read).toBeLessThan(compute);
-    expect(compute).toBeLessThan(validate);
-    expect(validate).toBeLessThan(transaction);
-    expect(transaction).toBeLessThan(write);
-    const writeHelper = managementSource.indexOf(
-      'export async function writeTopologyConfigMutation',
-    );
-    expect(writeHelper).toBeGreaterThan(-1);
-    const writer = managementSource.slice(writeHelper);
-    expect(writer).toContain('transaction: PSqlTransactionSql');
-    expect(writer).not.toContain('.begin(');
-    expect(appInboxSource.slice(read, write)).not.toContain('.begin(');
-  });
-
-  it('resolves server defaults, durable config, temporary override, and request options', () => {
-    const durable = {
-      groupRef: createGroupRef(),
-      config: {
-        topologyKind: 'tree' as const,
-        degreeLimit: 4,
-        treeMinSize: 6,
-        meshMinSize: 16,
-        meshParamK: 2,
-      },
-      version: 1,
-      createdAtEpochMs: 1,
-      updatedAtEpochMs: 1,
-      updatedByPrincipalId: 'owner',
-      requestId: 'durable-config',
-    };
-    const temporary = {
-      ...durable,
-      config: {
-        topologyKind: 'mesh' as const,
-        degreeLimit: 4,
-        treeMinSize: 6,
-        meshMinSize: 20,
-        meshParamK: 2,
-      },
-      version: 2,
-      expiresAtEpochMs: 10_000,
-    };
-
-    const view = resolveGroupTopologyConfig({
-      serverOptions: {
-        degreeLimit: 5,
-        treeMinSize: 6,
-        meshMinSize: 16,
-        meshParamK: 2,
-      },
-      durable,
-      temporary,
-      requestOptions: {
-        degreeLimit: 8,
-      },
-    });
-
-    expect(view.serverDefaults).toEqual({
-      topologyKind: 'auto',
-      degreeLimit: 5,
-      treeMinSize: 6,
-      meshMinSize: 16,
-      meshParamK: 2,
-    });
-    expect(view.effective).toEqual({
-      topologyKind: 'mesh',
-      degreeLimit: 8,
-      treeMinSize: 6,
-      meshMinSize: 20,
-      meshParamK: 2,
-    });
-    expect(view.durable).toEqual(durable);
-    expect(view.temporary).toEqual(temporary);
-    expect(view.requestOptions).toEqual({ degreeLimit: 8 });
-  });
-
-  it('defaults server config to auto topology plus threshold defaults', () => {
-    expect(readDefaultGroupTopologyConfig({})).toEqual({
-      topologyKind: 'auto',
-      degreeLimit: 5,
-      treeMinSize: 5,
-      meshMinSize: 16,
-      meshParamK: 2,
-    });
-  });
-
-  it('rejects invalid topology config patches and effective config combinations', () => {
-    expect(() => validateGroupTopologyConfigPatch({ degreeLimit: 0 })).toThrow(
-      GroupTopologyConfigValidationError,
-    );
-    expect(() => validateGroupTopologyConfigPatch({ treeMinSize: -1 })).toThrow(
-      GroupTopologyConfigValidationError,
-    );
-    expect(() => validateGroupTopologyConfigPatch({ meshMinSize: 1.5 })).toThrow(
-      GroupTopologyConfigValidationError,
-    );
-
-    expect(() =>
-      validateEffectiveGroupTopologyConfig({
-        topologyKind: 'auto',
-        degreeLimit: 5,
-        treeMinSize: 10,
-        meshMinSize: 9,
-        meshParamK: 2,
-      }),
-    ).toThrow(GroupTopologyConfigValidationError);
-    expect(() =>
-      validateEffectiveGroupTopologyConfig({
-        topologyKind: 'auto',
-        degreeLimit: 3,
-        treeMinSize: 5,
-        meshMinSize: 16,
-        meshParamK: 4,
-      }),
-    ).toThrow(GroupTopologyConfigValidationError);
-  });
-
-  it('defaults temporary override expiry to 15 minutes and caps it at 24 hours', () => {
-    expect(resolveOverrideExpiresAtEpochMs({ nowEpochMs: 1_000 })).toBe(
-      1_000 + DEFAULT_GROUP_TOPOLOGY_OVERRIDE_TTL_MS,
-    );
-    expect(
-      resolveOverrideExpiresAtEpochMs({
-        nowEpochMs: 1_000,
-        ttlMs: 5_000,
-      }),
-    ).toBe(6_000);
-    expect(
-      resolveOverrideExpiresAtEpochMs({
-        nowEpochMs: 1_000,
-        expiresAtEpochMs: 1_000 + MAX_GROUP_TOPOLOGY_OVERRIDE_TTL_MS + 10_000,
-      }),
-    ).toBe(1_000 + MAX_GROUP_TOPOLOGY_OVERRIDE_TTL_MS);
-  });
-
-  it('rejects temporary override expiries that are not in the future', () => {
-    expect(() =>
-      resolveOverrideExpiresAtEpochMs({
-        nowEpochMs: 1_000,
-        ttlMs: 0,
-      }),
-    ).toThrow(GroupTopologyConfigValidationError);
-    expect(() =>
-      resolveOverrideExpiresAtEpochMs({
-        nowEpochMs: 1_000,
-        expiresAtEpochMs: 999,
-      }),
-    ).toThrow(GroupTopologyConfigValidationError);
-  });
-});
-
-function createGroupRef() {
-  return {
-    applicationId: 'app-1',
-    workspaceId: 'workspace-1',
-    groupId: 'room-1',
-  };
-}
-function createConfigMutationInput(
-  input: Readonly<{
-    operation: 'putConfig' | 'putOverride';
-    config: Parameters<typeof validateGroupTopologyConfigPatch>[0];
-    durableDegreeLimit: number;
-    overrideDegreeLimit: number | null;
-  }>,
-) {
-  const groupRef = createGroupRef();
-  const stored = (degreeLimit: number, requestId: string) => ({
-    groupRef,
-    config: {
-      topologyKind: 'auto' as const,
-      degreeLimit,
-      treeMinSize: 5,
-      meshMinSize: 16,
-      meshParamK: 2,
-    },
-    version: 1,
-    createdAtEpochMs: 500,
-    updatedAtEpochMs: 500,
-    updatedByPrincipalId: 'owner',
-    requestId,
-  });
-  const durable = stored(input.durableDegreeLimit, 'durable');
-  const override =
-    input.overrideDegreeLimit === null
-      ? null
-      : {
-          ...stored(input.overrideDegreeLimit, 'override'),
-          expiresAtEpochMs: 10_000,
-        };
-  return {
-    command: {
-      operation: input.operation,
-      aggregateRef: groupRef,
-      commandId: `clear-${input.operation}`,
-      requestId: `clear-${input.operation}`,
-      input: {
-        config: input.config,
-        updatedByPrincipalId: 'owner',
-        ttlMs: input.operation === 'putOverride' ? 5_000 : null,
-        expiresAtEpochMs: null,
-      },
-    },
-    read: {
-      config: {
-        key: 'config',
-        value: durable,
-        entry: {
-          key: 'config',
-          value: JSON.stringify(durable),
-          expireAtTimestamp: Number.MAX_SAFE_INTEGER,
-          updatedTimestamp: new Date(0).toISOString(),
-          revision: 0,
-        },
-      },
-      override:
-        override === null
-          ? null
-          : {
-              key: 'override',
-              value: override,
-              entry: {
-                key: 'override',
-                value: JSON.stringify(override),
-                expireAtTimestamp: override.expiresAtEpochMs,
-                updatedTimestamp: new Date(0).toISOString(),
-                revision: 0,
-              },
-            },
-      configGeneration: null,
-      overrideGeneration: null,
-      invariantGeneration: null,
-      idempotency: null,
-      groupSnapshot: createGroupSnapshot(),
-      groupAuthorityGuard: createGroupAuthorityGuard(),
-    },
-    facts: {
-      requestedAtEpochMs: 1_000,
-      policyNowEpochMs: 1_000,
-      commandHash: `sha256:${'c'.repeat(64)}`,
-      attemptCount: 1,
-      isPlatformAdmin: false,
-      resolvedOverrideExpiresAtEpochMs: input.operation === 'putOverride' ? 6_000 : null,
-    },
-    serverDefaults: {},
-  } as const;
-}
-
-function createGroupSnapshot(): GroupSnapshot {
-  const groupRef = createGroupRef();
-  return {
-    stateRevision: 1,
-    causalRevision: { groupRevision: 1, presenceRevision: 0 },
-    group: {
-      ...groupRef,
-      slug: null,
-      displayName: 'Room 1',
-      description: null,
-      kind: 'room' as const,
-      status: 'active' as const,
-      archived: null,
-      deleted: null,
-      joinMode: 'open' as const,
-      maxMembers: null,
-      maxSessionsPerMember: null,
-      metadata: {},
-      snapshotVersion: 1,
-      metadataVersion: 0,
-      rosterVersion: 1,
-      presenceVersion: 0,
-      activeMemberCount: 1,
-      ownerPrincipalId: 'owner',
-      expiresAtEpochMs: null,
-      emptySinceEpochMs: null,
-      purgeAfterEpochMs: null,
-      created: {
-        atEpochMs: 1,
-        actor: { kind: 'principal', principalId: 'owner' },
-        reason: null,
-        traceId: null,
-        requestId: null,
-      },
-      updated: {
-        atEpochMs: 1,
-        actor: { kind: 'principal', principalId: 'owner' },
-        reason: null,
-        traceId: null,
-        requestId: null,
-      },
-    },
-    members: [
-      {
-        ...groupRef,
-        principalId: 'owner',
-        role: 'owner' as const,
-        status: 'active' as const,
-        invitedByPrincipalId: null,
-        invitationExpiresAtEpochMs: null,
-        left: null,
-        removed: null,
-        banned: null,
-        joined: {
-          atEpochMs: 1,
-          actor: { kind: 'principal', principalId: 'owner' },
-          reason: null,
-          traceId: null,
-          requestId: null,
-        },
-        updated: {
-          atEpochMs: 1,
-          actor: { kind: 'principal', principalId: 'owner' },
-          reason: null,
-          traceId: null,
-          requestId: null,
-        },
-      },
-    ],
-    activeSessions: [],
-    memberCount: 1,
-    onlineMemberCount: 0,
-  };
-}
-
-function createGroupAuthorityGuard(revision = 0) {
-  const group = createGroupSnapshot().group;
-  return {
-    groupRef: createGroupRef(),
-    causalGroupRevision: 1,
-    entry: {
-      key: 'group-authority',
-      value: JSON.stringify(group),
-      expireAtTimestamp: Number.MAX_SAFE_INTEGER,
-      updatedTimestamp: new Date(0).toISOString(),
-      revision,
-    },
-  };
-}
-
-function deepFreeze<T>(value: T): T {
-  if (value && typeof value === 'object') {
-    Object.freeze(value);
-    for (const nested of Object.values(value as Record<string, unknown>)) {
-      deepFreeze(nested);
-    }
-  }
-  return value;
-}
