+++ plans/repo-style-lineages/rallar-group-topology-server-pr-a-provenance.md
@@ -1,0 +1,79 @@
+# Rallar group topology server PR-A lineage provenance
+
+Base: `8b1ebf542d12c05a5ac226d3d07e543a171a2626`
+
+This evidence grants no blanket debt allowance. Semantically new code receives zero historical capacity.
+Each row below identifies only mechanically
+source-derived behavior. Changed contracts, branches, validators, helpers, and
+tests are reviewed as new code.
+
+## Mutation deterministic-validation entry
+
+Source: `packages/shared-server/rallar-system/services/group-topology-config-mutations.ts@c3ff5865c14de0df94f53468f20faacfa2021eda`
+
+Source symbol: `validateTopologyConfigMutation`
+
+Source span: lines 203-249 at the exact base
+
+Target: `packages/shared-server/rallar-system/topology/config/mutation/validate-topology-config-mutation.ts`
+
+Target blob: `c2928542681065dacfb28968c5ca3bd98f2273b0`
+
+Target symbol: `validateTopologyConfigMutation`
+
+Target span: the complete exported function at the candidate tree
+
+Magnitude: one deterministic JSON recomputation comparison, one receipt-hash
+comparison, one mutation-record validation, and one outbox identity comparison;
+no inherited warning capacity is claimed.
+
+Derivation: the target preserves the source phase and error order while named
+input and subordinate receipt/record owners replace anonymous multi-field
+contracts. The row is traceability evidence only; refactored lines are ineligible
+for source warning capacity.
+
+## Config defaults and resolution owner
+
+Source: `packages/shared-server/rallar-system/services/group-topology-config-service.ts@a78dc11667c80be903b486049ce58a4734334017`
+
+Source symbol: `resolveGroupTopologyConfig`
+
+Source span: lines 123-160 at the exact base
+
+Target: `packages/shared-server/rallar-system/topology/config/group-topology-config.ts`
+
+Target blob: `45509981e74365c6b011932b067c2bb44abf177f`
+
+Target symbol: `resolveGroupTopologyConfig`
+
+Target span: the complete exported function at the candidate tree
+
+Magnitude: one pure defaults/precedence/resolution owner; zero inherited
+construction, layout, output-contract, or unknown-boundary findings.
+
+Derivation: path relocation plus formatter-only layout and the required relative
+import adjustment. The validation order, spreads, clone points, property order,
+and error owner remain source-derived.
+
+## Shared topology authority proof
+
+Source: `packages/shared-server/rallar-system/services/topology-mutation-authority-proof.ts@2be25a4d4b071c85fd6867842590adf38738e5e6`
+
+Source symbol: `createTopologyMutationAuthorityProof`
+
+Source span: lines 16-48 at the exact base
+
+Target: `packages/shared-server/rallar-system/topology/inbox/topology-mutation-authority-proof.ts`
+
+Target blob: `8f95077ffe5abbbbc02a09216f49736a9f175f89`
+
+Target symbol: `createTopologyMutationAuthorityProof`
+
+Target span: the complete exported function at the candidate tree
+
+Magnitude: one proof contract and one HMAC constructor; zero inherited warning
+capacity.
+
+Derivation: path relocation and relative import adjustment only. Proof property
+order, HMAC purpose, key algorithm, byte encoding, and hexadecimal projection
+remain source-derived.
