+++ plans/repo-style-lineages/rallar-group-topology-server-pr-a.json
@@ -1,0 +1,33 @@
+{
+  "version": 1,
+  "lineages": [
+    {
+      "mergeBase": "8b1ebf542d12c05a5ac226d3d07e543a171a2626",
+      "source": {
+        "path": "packages/shared-server/rallar-system/services/group-topology-config-mutations.ts",
+        "blob": "c3ff5865c14de0df94f53468f20faacfa2021eda"
+      },
+      "targets": [
+        "packages/shared-server/rallar-system/topology/config/mutation/validate-topology-config-mutation.ts"
+      ]
+    },
+    {
+      "mergeBase": "8b1ebf542d12c05a5ac226d3d07e543a171a2626",
+      "source": {
+        "path": "packages/shared-server/rallar-system/services/group-topology-config-service.ts",
+        "blob": "a78dc11667c80be903b486049ce58a4734334017"
+      },
+      "targets": ["packages/shared-server/rallar-system/topology/config/group-topology-config.ts"]
+    },
+    {
+      "mergeBase": "8b1ebf542d12c05a5ac226d3d07e543a171a2626",
+      "source": {
+        "path": "packages/shared-server/rallar-system/services/topology-mutation-authority-proof.ts",
+        "blob": "2be25a4d4b071c85fd6867842590adf38738e5e6"
+      },
+      "targets": [
+        "packages/shared-server/rallar-system/topology/inbox/topology-mutation-authority-proof.ts"
+      ]
+    }
+  ]
+}
